Moved 3 components to one folder. Installed fontawesome. Updated App.jsx. Changed text color for title on website list.
