+++ frontend/package.json
@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@fortawesome/fontawesome-free": "^6.6.0",
     "axios": "^1.7.7",
     "bootstrap": "^5.3.3",
     "cra-template": "1.2.0",
